Discord - separated datamodels
