+++ src/Artemis.Plugins.Modules.Discord/DataModels/DiscordDataModel.cs
@@ -13,127 +13,4 @@ namespace Artemis.Plugins.Modules.Discord.DataModels
 
         public DataModelEvent<DiscordNotificationEventArgs> Notification { get; set; } = new DataModelEvent<DiscordNotificationEventArgs>();
     }
-
-    public class DiscordNotificationEventArgs : DataModelEventArgs
-    {
-        public string ChannelId { get; set; }
-        public User Author { get; set; }
-        public string IconUrl { get; set; }
-        public string Title { get; set; }
-        public string Body { get; set; }
-    }
-
-    public class DiscordVoiceSettingsDataModel : DataModel
-    {
-        public bool Muted { get; set; }
-        public bool Deafened { get; set; }
-        public bool AutomaticGainControl { get; set; }
-        public bool EchoCancellation { get; set; }
-        public bool NoiseSuppression { get; set; }
-        public bool Qos { get; set; }
-        public bool SilenceWarning { get; set; }
-        public DiscordVoiceMode Mode { get; set; } = new DiscordVoiceMode();
-    }
-
-    public enum DiscordVoiceModeType
-    {
-        PUSH_TO_TALK,
-        VOICE_ACTIVITY
-    }
-
-    public class DiscordVoiceMode : DataModel
-    {
-        public DiscordVoiceModeType Type { get; set; }
-        public bool AutoThreshold { get; set; }
-        public float Threshold { get; set; }
-        public DiscordShortcut[] Shortcut { get; set; } = System.Array.Empty<DiscordShortcut>();
-        public float Delay { get; set; }
-    }
-
-    public enum DiscordKeyType : int
-    {
-        KEYBOARD_KEY,
-        MOUSE_BUTTON,
-        KEYBOARD_MODIFIER_KEY,
-        GAMEPAD_BUTTON
-    }
-
-    public class DiscordShortcut : DataModel
-    {
-        public DiscordKeyType Type { get; set; }
-        public int Code { get; set; }
-        public string Name { get; set; }
-    }
-
-    public class DiscordUserDataModel : DataModel
-    {
-        public string Username { get; set; }
-
-        public string Discriminator { get; set; }
-
-        public string Id { get; set; }
-    }
-
-    public class DiscordVoiceConnectionStatusDataModel : DataModel
-    {
-        public DiscordVoiceChannelState State { get; set; }
-        public string Hostname { get; set; }
-        public float? Ping { get; set; }
-        public bool Speaking { get; set; }
-        public bool IsVoiceConnected => State == DiscordVoiceChannelState.VOICE_CONNECTED;
-        public DataModelEvent Connected { get; set; } = new DataModelEvent();
-        public DataModelEvent Disconnected { get; set; } = new DataModelEvent();
-    }
-
-    public enum DiscordVoiceChannelState
-    {
-        /// <summary>
-        /// TCP disconnected
-        /// </summary>
-        DISCONNECTED,
-        /// <summary>
-        /// Waiting for voice endpoint
-        /// </summary>
-        AWAITING_ENDPOINT,
-
-        /// <summary>
-        /// TCP authenticating
-        /// </summary>
-        AUTHENTICATING,
-
-        /// <summary>
-        /// TCP connecting
-        /// </summary>
-        CONNECTING,//TCP connecting
-
-        /// <summary>
-        /// TCP connected
-        /// </summary>
-        CONNECTED,
-
-        /// <summary>
-        /// TCP connected, Voice disconnected
-        /// </summary>
-        VOICE_DISCONNECTED,
-
-        /// <summary>
-        /// TCP connected, Voice connecting
-        /// </summary>
-        VOICE_CONNECTING,
-
-        /// <summary>
-        /// TCP connected, Voice connected
-        /// </summary>
-        VOICE_CONNECTED,
-
-        /// <summary>
-        /// No route to host
-        /// </summary>
-        NO_ROUTE,
-
-        /// <summary>
-        /// WebRTC ice checking
-        /// </summary>
-        ICE_CHECKING,
-    }
 }
