+++ src/Artemis.Plugins.Modules.Discord/DataModels/DiscordKeyType.cs
@@ -1,0 +1,10 @@
+﻿namespace Artemis.Plugins.Modules.Discord.DataModels
+{
+    public enum DiscordKeyType
+    {
+        KEYBOARD_KEY,
+        MOUSE_BUTTON,
+        KEYBOARD_MODIFIER_KEY,
+        GAMEPAD_BUTTON
+    }
+}
