+++ src/Artemis.Plugins.Modules.Discord/DataModels/DiscordShortcut.cs
@@ -1,0 +1,11 @@
+﻿using Artemis.Core.Modules;
+
+namespace Artemis.Plugins.Modules.Discord.DataModels
+{
+    public class DiscordShortcut : DataModel
+    {
+        public DiscordKeyType Type { get; set; }
+        public int Code { get; set; }
+        public string Name { get; set; }
+    }
+}
