+++ src/Artemis.Plugins.Modules.Discord/DataModels/DiscordVoiceModeType.cs
@@ -1,0 +1,8 @@
+﻿namespace Artemis.Plugins.Modules.Discord.DataModels
+{
+    public enum DiscordVoiceModeType
+    {
+        PUSH_TO_TALK,
+        VOICE_ACTIVITY
+    }
+}
