--- src/Artemis.Plugins.Modules.Discord/DiscordPackets/CommandData/SpeakingStart.cs
@@ -1,4 +1,0 @@
-﻿namespace Artemis.Plugins.Modules.Discord
-{
-    public record SpeakingStart(string UserId);
-}
