+++ src/Artemis.Plugins.Modules.Discord/DiscordPackets/CommandData/SpeakingStartStop.cs
@@ -1,0 +1,4 @@
+﻿namespace Artemis.Plugins.Modules.Discord
+{
+    public record SpeakingStartStop(string UserId);
+}
