--- src/Artemis.Plugins.Modules.Discord/DiscordPackets/CommandData/SpeakingStop.cs
@@ -1,4 +1,0 @@
-﻿namespace Artemis.Plugins.Modules.Discord
-{
-    public record SpeakingStop(string UserId);
-}
