+++ src/Artemis.Plugins.Modules.Discord/DiscordPackets/CommandData/Subscription.cs
@@ -2,7 +2,7 @@
 
 namespace Artemis.Plugins.Modules.Discord
 {
-    public record Subscribe(
+    public record Subscription(
         [JsonProperty("evt")]
         string Event
     );
